Adding flags to most HAL methods and extending existing ones to i64. (#20368)
**This is a breaking HAL change and will require recompilation of
VMFBs.** There _should_ be a useful error indicating that emitted if
attempting to load an existing VMFB.

There's been TODOs in the HAL imports for awhile waiting for a breakage
and #20240 does that. To prevent more PRs like that this adds flag
arguments to most HAL methods so that we can add new flags in the
future. Some existing flags were i32 and have been changed to i64 for
consistency.

Of note:
* `hal.device.queue.execute` split to `hal.device.queue.barrier` instead
of relying on empty command buffer list
* `hal.command_buffer.advise_buffer` added, but not yet in IR
* `hal.command_buffer.fill_buffer` max pattern length changed to i64 to
match `hal.device.queue.fill`

+++ runtime/src/iree/hal/channel.h
@@ -27,7 +27,7 @@ typedef struct iree_hal_device_t iree_hal_device_t;
 enum iree_hal_channel_flag_bits_t {
   IREE_HAL_CHANNEL_FLAG_NONE = 0u,
 };
-typedef uint32_t iree_hal_channel_flags_t;
+typedef uint64_t iree_hal_channel_flags_t;
 
 // Specifies that the channel should use environment settings if available.
 #define IREE_HAL_CHANNEL_RANK_DEFAULT ((int32_t) - 1)
